requirements.txt
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.1
fonttools==4.39.4
iniconfig==2.0.0
kiwisolver==1.4.4
loguru==0.7.0
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pytest==7.3.1
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.10.1
six==1.16.0
tomli==2.0.1
